feat: Added fuzzers logic through formats for autointegrating with commands and exploits
Migrated from manual threads to async event loops for faster development in an async context, the necessities came from the fuzzer we need somehow to queue several calls to commands with different payloads so we used the python async capabilities for that, why not? threading? because seems like a cannot to kill a mosquito (a.k.a. laziness and pragmacity)